Pinned image for Traefik to use `traefik:v1.7`
This resolves "command traefik error: field not found, node: redirect" errors that put Traefik into an endless restart loop.
See containous/traefik#5119 for changes causing the BC break; this is a problem since Warden currently leverages the removed configuration settings in the TOML file that's generated and passed into the Traefik container.

Thanks to @magedevjosh for finding and reporting this issue!