Disabling by-default logging of DNS queries in dnsmasq container; see below for why
On macOS a file at /etc/resolver/test is used to route only queries for *.test to the dnsmasq container. This works and only queries for .test domains are routed (and previously were logged) by the dnsmasq container. On Linux systems however, the only way to setup on name resolution for *.test domains is to configure the network manager to send all queries through the dnsmasq container. As @talesh pointed out, this poses a potential privacy issue for those linux users among us. This mitigates that by disabling the logging by default. Logging may still be enabled by going to https://dnsmasq.warden.test/ in a browser, and altering the loaded configuration.