Remove deprecated first-class commands down, restart, start, stop, and up
From 0.6.0 release notes:

* Added `warden svc` command to control global services replacing `warden start`, `warden stop`, `warden up`, `warden down`, and `warden restart` and offering further flexibility as this works similar to warden env in that any verb known to docker-compose may be used in orchestrating global services such as traefik, dnsmasq and portainer; for example, warden svc up does what warden up did previously.

When this was done, these old commands were kept and printed a warning such as the following:

> WARNING: This command is deprecated as of 0.6.0 and will be removed in 0.7.0; please use 'warden svc start' instead.

These commands were for various reasons not cleaned up in the 0.7.0 release as originally planned. This removes them now requiring the use of `warden svc <verb>` to control global services.