As of Mutagen 0.15.0, It displays "Connected: Yes" for beta and alpha if they are connected
I have not been able to identify where "Connected state: Connected" used to be as of yet, but I'm making the assumption it changed to "Connected:" when it first appears in list_monitor_common.go, which is v0.15.0.

Prior to this commit, Warden would always run `warden sync start` after `warden env up` - which for larger projects on machines with Antivirus is.. incredibly annoying.